Add files via upload
Changed the JS listener to mouseenter and mouseleave and inserted the company dropdown content inside the company navlink div. This is to make sure that the dropdown stays active even when the mouse leaves the company navlink and onto the company dropdown content, because the company dropdown content is now part of the company navlink component.